requirements.txt
Django==1.8.5
MySQL-python==1.2.5
Unipath==1.1
django-autoslug==1.9.3
django-formtools==1.0